Add DayPickerSingleDatePicker component
