+++ .storybook/config.js
@@ -59,6 +59,7 @@ function loadStories() {
   require('../stories/SingleDatePicker_calendar');
   require('../stories/SingleDatePicker_day');
   require('../stories/DayPickerRangeController');
+  require('../stories/DayPickerSingleDateController');
   require('../stories/DayPicker');
 }
