+++ src/components/SingleDatePicker.jsx
@@ -4,7 +4,6 @@ import cx from 'classnames';
 import Portal from 'react-portal';
 import { forbidExtraProps } from 'airbnb-prop-types';
 import { addEventListener, removeEventListener } from 'consolidated-events';
-import values from 'object.values';
 import isTouchDevice from 'is-touch-device';
 
 import SingleDatePickerShape from '../shapes/SingleDatePickerShape';
@@ -15,25 +14,18 @@ import toMomentObject from '../utils/toMomentObject';
 import toLocalizedDateString from '../utils/toLocalizedDateString';
 import getResponsiveContainerStyles from '../utils/getResponsiveContainerStyles';
 
-import getVisibleDays from '../utils/getVisibleDays';
-import isDayVisible from '../utils/isDayVisible';
-
 import toISODateString from '../utils/toISODateString';
-import toISOMonthString from '../utils/toISOMonthString';
 
 import SingleDatePickerInput from './SingleDatePickerInput';
-import DayPicker from './DayPicker';
+import DayPickerSingleDateController from './DayPickerSingleDateController';
 
 import CloseButton from '../svg/close.svg';
 
 import isInclusivelyAfterDay from '../utils/isInclusivelyAfterDay';
-import isSameDay from '../utils/isSameDay';
-import isAfterDay from '../utils/isAfterDay';
 
 import {
   HORIZONTAL_ORIENTATION,
   VERTICAL_ORIENTATION,
-  VERTICAL_SCROLLABLE,
   ANCHOR_LEFT,
   ANCHOR_RIGHT,
   DAY_SIZE,
@@ -102,46 +94,20 @@ export default class SingleDatePicker extends React.Component {
     super(props);
 
     this.isTouchDevice = false;
-    this.today = moment();
-
-    this.modifiers = {
-      today: day => this.isToday(day),
-      blocked: day => this.isBlocked(day),
-      'blocked-calendar': day => props.isDayBlocked(day),
-      'blocked-out-of-range': day => props.isOutsideRange(day),
-      'highlighted-calendar': day => props.isDayHighlighted(day),
-      valid: day => !this.isBlocked(day),
-      hovered: day => this.isHovered(day),
-      selected: day => this.isSelected(day),
-    };
-
-    const { currentMonth, visibleDays } = this.getStateForNewMonth(props);
 
     this.state = {
       dayPickerContainerStyles: {},
-      hoverDate: null,
       isDayPickerFocused: false,
       isInputFocused: false,
-      currentMonth,
-      visibleDays,
     };
-
-    this.onDayMouseEnter = this.onDayMouseEnter.bind(this);
-    this.onDayMouseLeave = this.onDayMouseLeave.bind(this);
-    this.onDayClick = this.onDayClick.bind(this);
 
     this.onDayPickerFocus = this.onDayPickerFocus.bind(this);
     this.onDayPickerBlur = this.onDayPickerBlur.bind(this);
-
-    this.onPrevMonthClick = this.onPrevMonthClick.bind(this);
-    this.onNextMonthClick = this.onNextMonthClick.bind(this);
 
     this.onChange = this.onChange.bind(this);
     this.onFocus = this.onFocus.bind(this);
     this.onClearFocus = this.onClearFocus.bind(this);
     this.clearDate = this.clearDate.bind(this);
-
-    this.getFirstFocusableDay = this.getFirstFocusableDay.bind(this);
 
     this.responsivizePickerPosition = this.responsivizePickerPosition.bind(this);
   }
@@ -152,7 +118,7 @@ export default class SingleDatePicker extends React.Component {
       window,
       'resize',
       this.responsivizePickerPosition,
-      { passive: true },
+    { passive: true },
     );
     this.responsivizePickerPosition();
 
@@ -163,107 +129,6 @@ export default class SingleDatePicker extends React.Component {
     }
 
     this.isTouchDevice = isTouchDevice();
-  }
-
-  componentWillReceiveProps(nextProps) {
-    const {
-      date,
-      focused,
-      isOutsideRange,
-      isDayBlocked,
-      isDayHighlighted,
-      initialVisibleMonth,
-      numberOfMonths,
-      enableOutsideDays,
-    } = nextProps;
-    let { visibleDays } = this.state;
-
-    if (isOutsideRange !== this.props.isOutsideRange) {
-      this.modifiers['blocked-out-of-range'] = day => isOutsideRange(day);
-    }
-
-    if (isDayBlocked !== this.props.isDayBlocked) {
-      this.modifiers['blocked-calendar'] = day => isDayBlocked(day);
-    }
-
-    if (isDayHighlighted !== this.props.isDayHighlighted) {
-      this.modifiers['highlighted-calendar'] = day => isDayHighlighted(day);
-    }
-
-    if (
-      initialVisibleMonth !== this.props.initialVisibleMonth ||
-      numberOfMonths !== this.props.numberOfMonths ||
-      enableOutsideDays !== this.props.enableOutsideDays
-    ) {
-      const newMonthState = this.getStateForNewMonth(nextProps);
-      const currentMonth = newMonthState.currentMonth;
-      visibleDays = newMonthState.visibleDays;
-      this.setState({
-        currentMonth,
-        visibleDays,
-      });
-    }
-
-    const didDateChange = date !== this.props.date;
-    const didFocusChange = focused !== this.props.focused;
-
-    let modifiers = {};
-
-    if (didDateChange) {
-      modifiers = this.deleteModifier(modifiers, this.props.date, 'selected');
-      modifiers = this.addModifier(modifiers, date, 'selected');
-    }
-
-    if (didFocusChange) {
-      values(visibleDays).forEach((days) => {
-        Object.keys(days).forEach((day) => {
-          const momentObj = moment(day);
-          if (this.isBlocked(momentObj)) {
-            modifiers = this.addModifier(modifiers, momentObj, 'blocked');
-          } else {
-            modifiers = this.deleteModifier(modifiers, momentObj, 'blocked');
-          }
-
-          if (isOutsideRange(momentObj)) {
-            modifiers = this.addModifier(modifiers, momentObj, 'blocked-out-of-range');
-          } else {
-            modifiers = this.deleteModifier(modifiers, momentObj, 'blocked-out-of-range');
-          }
-
-          if (isDayBlocked(momentObj)) {
-            modifiers = this.addModifier(modifiers, momentObj, 'blocked-calendar');
-          } else {
-            modifiers = this.deleteModifier(modifiers, momentObj, 'blocked-calendar');
-          }
-
-          if (isDayHighlighted(momentObj)) {
-            modifiers = this.addModifier(modifiers, momentObj, 'highlighted-calendar');
-          } else {
-            modifiers = this.deleteModifier(modifiers, momentObj, 'highlighted-calendar');
-          }
-        });
-      });
-    }
-
-    const today = moment();
-    if (!isSameDay(this.today, today)) {
-      modifiers = this.deleteModifier(modifiers, this.today, 'today');
-      modifiers = this.addModifier(modifiers, today, 'today');
-      this.today = today;
-    }
-
-    if (Object.keys(modifiers).length > 0) {
-      this.setState({
-        visibleDays: {
-          ...visibleDays,
-          ...modifiers,
-        },
-      });
-    }
-  }
-
-  componentWillUpdate() {
-    this.today = moment();
   }
 
   componentDidUpdate(prevProps) {
@@ -279,75 +144,24 @@ export default class SingleDatePicker extends React.Component {
 
   onChange(dateString) {
     const {
-      startDate,
       isOutsideRange,
       keepOpenOnDateSelect,
       onDateChange,
       onFocusChange,
       onClose,
     } = this.props;
-    const endDate = toMomentObject(dateString, this.getDisplayFormat());
+    const newDate = toMomentObject(dateString, this.getDisplayFormat());
 
-    const isValid = endDate && !isOutsideRange(endDate);
+    const isValid = newDate && !isOutsideRange(newDate);
     if (isValid) {
-      onDateChange(endDate);
+      onDateChange(newDate);
       if (!keepOpenOnDateSelect) {
         onFocusChange({ focused: false });
-        onClose({ startDate, endDate });
+        onClose({ date: newDate });
       }
     } else {
       onDateChange(null);
     }
-  }
-
-  onDayClick(day, e) {
-    if (e) e.preventDefault();
-    if (this.isBlocked(day)) return;
-    const {
-      onDateChange,
-      keepOpenOnDateSelect,
-      onFocusChange,
-      onClose,
-      startDate,
-      endDate,
-    } = this.props;
-
-    onDateChange(day);
-    if (!keepOpenOnDateSelect) {
-      onFocusChange({ focused: null });
-      onClose({ startDate, endDate });
-    }
-  }
-
-  onDayMouseEnter(day) {
-    if (this.isTouchDevice) return;
-    const { hoverDate, visibleDays } = this.state;
-
-    let modifiers = this.deleteModifier({}, hoverDate, 'hovered');
-    modifiers = this.addModifier(modifiers, day, 'hovered');
-
-    this.setState({
-      hoverDate: day,
-      visibleDays: {
-        ...visibleDays,
-        ...modifiers,
-      },
-    });
-  }
-
-  onDayMouseLeave() {
-    const { hoverDate, visibleDays } = this.state;
-    if (this.isTouchDevice || !hoverDate) return;
-
-    const modifiers = this.deleteModifier({}, hoverDate, 'hovered');
-
-    this.setState({
-      hoverDate: null,
-      visibleDays: {
-        ...visibleDays,
-        ...modifiers,
-      },
-    });
   }
 
   onFocus() {
@@ -392,52 +206,6 @@ export default class SingleDatePicker extends React.Component {
     });
   }
 
-  onPrevMonthClick() {
-    const { onPrevMonthClick, numberOfMonths, enableOutsideDays } = this.props;
-    const { currentMonth, visibleDays } = this.state;
-
-    const newVisibleDays = {};
-    Object.keys(visibleDays).sort().slice(0, numberOfMonths + 1).forEach((month) => {
-      newVisibleDays[month] = visibleDays[month];
-    });
-
-    const prevMonth = currentMonth.clone().subtract(1, 'month');
-    const prevMonthVisibleDays = getVisibleDays(prevMonth, 1, enableOutsideDays);
-
-    this.setState({
-      currentMonth: prevMonth,
-      visibleDays: {
-        ...newVisibleDays,
-        ...this.getModifiers(prevMonthVisibleDays),
-      },
-    });
-
-    onPrevMonthClick();
-  }
-
-  onNextMonthClick() {
-    const { onNextMonthClick, numberOfMonths, enableOutsideDays } = this.props;
-    const { currentMonth, visibleDays } = this.state;
-
-    const newVisibleDays = {};
-    Object.keys(visibleDays).sort().slice(1).forEach((month) => {
-      newVisibleDays[month] = visibleDays[month];
-    });
-
-    const nextMonth = currentMonth.clone().add(numberOfMonths, 'month');
-    const nextMonthVisibleDays = getVisibleDays(nextMonth, 1, enableOutsideDays);
-
-    this.setState({
-      currentMonth: currentMonth.clone().add(1, 'month'),
-      visibleDays: {
-        ...newVisibleDays,
-        ...this.getModifiers(nextMonthVisibleDays),
-      },
-    });
-
-    onNextMonthClick();
-  }
-
   getDateString(date) {
     const displayFormat = this.getDisplayFormat();
     if (date && displayFormat) {
@@ -448,7 +216,6 @@ export default class SingleDatePicker extends React.Component {
 
   getDayPickerContainerClasses() {
     const { orientation, withPortal, withFullScreenPortal, anchorDirection, isRTL } = this.props;
-    const { hoverDate } = this.state;
 
     const dayPickerClassName = cx('SingleDatePicker__picker', {
       'SingleDatePicker__picker--direction-left': anchorDirection === ANCHOR_LEFT,
@@ -457,7 +224,6 @@ export default class SingleDatePicker extends React.Component {
       'SingleDatePicker__picker--vertical': orientation === VERTICAL_ORIENTATION,
       'SingleDatePicker__picker--portal': withPortal || withFullScreenPortal,
       'SingleDatePicker__picker--full-screen-portal': withFullScreenPortal,
-      'SingleDatePicker__picker--valid-date-hovered': hoverDate && !this.isBlocked(hoverDate),
       'SingleDatePicker__picker--rtl': isRTL,
     });
 
@@ -467,159 +233,6 @@ export default class SingleDatePicker extends React.Component {
   getDisplayFormat() {
     const { displayFormat } = this.props;
     return typeof displayFormat === 'string' ? displayFormat : displayFormat();
-  }
-
-  getFirstFocusableDay(newMonth) {
-    const { date, numberOfMonths } = this.props;
-
-    let focusedDate = newMonth.clone().startOf('month');
-    if (date) {
-      focusedDate = date.clone();
-    }
-
-    if (this.isBlocked(focusedDate)) {
-      const days = [];
-      const lastVisibleDay = newMonth.clone().add(numberOfMonths - 1, 'months').endOf('month');
-      let currentDay = focusedDate.clone();
-      while (!isAfterDay(currentDay, lastVisibleDay)) {
-        currentDay = currentDay.clone().add(1, 'day');
-        days.push(currentDay);
-      }
-
-      const viableDays = days.filter(day => !this.isBlocked(day) && isAfterDay(day, focusedDate));
-      if (viableDays.length > 0) focusedDate = viableDays[0];
-    }
-
-    return focusedDate;
-  }
-
-  getModifiers(visibleDays) {
-    const modifiers = {};
-    Object.keys(visibleDays).forEach((month) => {
-      modifiers[month] = {};
-      visibleDays[month].forEach((day) => {
-        modifiers[month][toISODateString(day)] = this.getModifiersForDay(day);
-      });
-    });
-
-    return modifiers;
-  }
-
-  getModifiersForDay(day) {
-    return new Set(Object.keys(this.modifiers).filter(modifier => this.modifiers[modifier](day)));
-  }
-
-  getStateForNewMonth(nextProps) {
-    const { initialVisibleMonth, date, numberOfMonths, enableOutsideDays } = nextProps;
-    const initialVisibleMonthThunk = initialVisibleMonth || (date ? () => date : () => this.today);
-    const currentMonth = initialVisibleMonthThunk();
-    const visibleDays =
-      this.getModifiers(getVisibleDays(currentMonth, numberOfMonths, enableOutsideDays));
-    return { currentMonth, visibleDays };
-  }
-
-  addModifier(updatedDays, day, modifier) {
-    const { numberOfMonths: numberOfVisibleMonths, enableOutsideDays, orientation } = this.props;
-    const { currentMonth: firstVisibleMonth, visibleDays } = this.state;
-
-    let currentMonth = firstVisibleMonth;
-    let numberOfMonths = numberOfVisibleMonths;
-    if (orientation !== VERTICAL_SCROLLABLE) {
-      currentMonth = currentMonth.clone().subtract(1, 'month');
-      numberOfMonths += 2;
-    }
-    if (!day || !isDayVisible(day, currentMonth, numberOfMonths, enableOutsideDays)) {
-      return updatedDays;
-    }
-
-    const iso = toISODateString(day);
-
-    let updatedDaysAfterAddition = { ...updatedDays };
-    if (enableOutsideDays) {
-      const monthsToUpdate = Object.keys(visibleDays).filter(monthKey => (
-        Object.keys(visibleDays[monthKey]).indexOf(iso) > -1
-      ));
-
-      updatedDaysAfterAddition = monthsToUpdate.reduce((days, monthIso) => {
-        const month = updatedDays[monthIso] || visibleDays[monthIso];
-        const modifiers = new Set(month[iso]);
-        modifiers.add(modifier);
-        return {
-          ...days,
-          [monthIso]: {
-            ...month,
-            [iso]: modifiers,
-          },
-        };
-      }, updatedDaysAfterAddition);
-    } else {
-      const monthIso = toISOMonthString(day);
-      const month = updatedDays[monthIso] || visibleDays[monthIso];
-
-      const modifiers = new Set(month[iso]);
-      modifiers.add(modifier);
-      updatedDaysAfterAddition = {
-        ...updatedDaysAfterAddition,
-        [monthIso]: {
-          ...month,
-          [iso]: modifiers,
-        },
-      };
-    }
-
-    return updatedDaysAfterAddition;
-  }
-
-  deleteModifier(updatedDays, day, modifier) {
-    const { numberOfMonths: numberOfVisibleMonths, enableOutsideDays, orientation } = this.props;
-    const { currentMonth: firstVisibleMonth, visibleDays } = this.state;
-
-    let currentMonth = firstVisibleMonth;
-    let numberOfMonths = numberOfVisibleMonths;
-    if (orientation !== VERTICAL_SCROLLABLE) {
-      currentMonth = currentMonth.clone().subtract(1, 'month');
-      numberOfMonths += 2;
-    }
-    if (!day || !isDayVisible(day, currentMonth, numberOfMonths, enableOutsideDays)) {
-      return updatedDays;
-    }
-
-    const iso = toISODateString(day);
-
-    let updatedDaysAfterDeletion = { ...updatedDays };
-    if (enableOutsideDays) {
-      const monthsToUpdate = Object.keys(visibleDays).filter(monthKey => (
-        Object.keys(visibleDays[monthKey]).indexOf(iso) > -1
-      ));
-
-      updatedDaysAfterDeletion = monthsToUpdate.reduce((days, monthIso) => {
-        const month = updatedDays[monthIso] || visibleDays[monthIso];
-        const modifiers = new Set(month[iso]);
-        modifiers.delete(modifier);
-        return {
-          ...days,
-          [monthIso]: {
-            ...month,
-            [iso]: modifiers,
-          },
-        };
-      }, updatedDaysAfterDeletion);
-    } else {
-      const monthIso = toISOMonthString(day);
-      const month = updatedDays[monthIso] || visibleDays[monthIso];
-
-      const modifiers = new Set(month[iso]);
-      modifiers.delete(modifier);
-      updatedDaysAfterDeletion = {
-        ...updatedDaysAfterDeletion,
-        [monthIso]: {
-          ...month,
-          [iso]: modifiers,
-        },
-      };
-    }
-
-    return updatedDaysAfterDeletion;
   }
 
   clearDate() {
@@ -664,24 +277,6 @@ export default class SingleDatePicker extends React.Component {
     }
   }
 
-  isBlocked(day) {
-    const { isDayBlocked, isOutsideRange } = this.props;
-    return isDayBlocked(day) || isOutsideRange(day);
-  }
-
-  isHovered(day) {
-    const { hoverDate } = this.state || {};
-    return isSameDay(day, hoverDate);
-  }
-
-  isSelected(day) {
-    return isSameDay(day, this.props.date);
-  }
-
-  isToday(day) {
-    return isSameDay(day, this.today);
-  }
-
   maybeRenderDayPickerWithPortal() {
     const { focused, withPortal, withFullScreenPortal } = this.props;
 
@@ -702,6 +297,10 @@ export default class SingleDatePicker extends React.Component {
 
   renderDayPicker() {
     const {
+      onDateChange,
+      date,
+      onFocusChange,
+      focused,
       enableOutsideDays,
       numberOfMonths,
       orientation,
@@ -710,7 +309,7 @@ export default class SingleDatePicker extends React.Component {
       navNext,
       withPortal,
       withFullScreenPortal,
-      focused,
+      initialVisibleMonth,
       renderMonth,
       renderDay,
       renderCalendarInfo,
@@ -719,8 +318,11 @@ export default class SingleDatePicker extends React.Component {
       phrases,
       daySize,
       isRTL,
+      isOutsideRange,
+      isDayBlocked,
+      isDayHighlighted,
     } = this.props;
-    const { dayPickerContainerStyles, isDayPickerFocused, currentMonth, visibleDays } = this.state;
+    const { dayPickerContainerStyles, isDayPickerFocused } = this.state;
 
     const onOutsideClick = (!withFullScreenPortal && withPortal) ? this.onClearFocus : undefined;
     const closeIcon = customCloseIcon || (<CloseButton />);
@@ -732,32 +334,30 @@ export default class SingleDatePicker extends React.Component {
         style={dayPickerContainerStyles}
         onClick={onOutsideClick}
       >
-        <DayPicker
+        <DayPickerSingleDateController
+          date={date}
+          onDateChange={onDateChange}
+          onFocusChange={onFocusChange}
           orientation={orientation}
           enableOutsideDays={enableOutsideDays}
-          modifiers={visibleDays}
           numberOfMonths={numberOfMonths}
-          onDayClick={this.onDayClick}
-          onDayMouseEnter={this.onDayMouseEnter}
-          onDayMouseLeave={this.onDayMouseLeave}
-          onPrevMonthClick={this.onPrevMonthClick}
-          onNextMonthClick={this.onNextMonthClick}
           monthFormat={monthFormat}
           withPortal={withPortal || withFullScreenPortal}
-          hidden={!focused}
+          focused={focused}
           hideKeyboardShortcutsPanel={hideKeyboardShortcutsPanel}
-          initialVisibleMonth={() => currentMonth}
+          initialVisibleMonth={initialVisibleMonth}
           navPrev={navPrev}
           navNext={navNext}
           renderMonth={renderMonth}
           renderDay={renderDay}
           renderCalendarInfo={renderCalendarInfo}
           isFocused={isDayPickerFocused}
-          getFirstFocusableDay={this.getFirstFocusableDay}
-          onBlur={this.onDayPickerBlur}
           phrases={phrases}
           daySize={daySize}
           isRTL={isRTL}
+          isOutsideRange={isOutsideRange}
+          isDayBlocked={isDayBlocked}
+          isDayHighlighted={isDayHighlighted}
         />
 
         {withFullScreenPortal && (
